stripe done

+++ database/migrations/2022_09_15_065020_create_realtor_payment_histories_table.php
@@ -17,7 +17,7 @@ class CreateRealtorPaymentHistoriesTable extends Migration
             $table->id();
             $table->integer('user_id');
             $table->string('subscription_id')->nullable();
-            $table->integer('amount');
+            $table->float('amount');
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP'));
         });
